# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Public discovery, inquiry, content, booking, and administrator API for From Kano to the World.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/tours:
    get:
      operationId: getAdminTours
      tags: [admin]
      summary: List all tours for admin
      responses:
        "200":
          description: All tours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tour"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createAdminTour
      tags: [admin]
      summary: Create a tour
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TourInput"
      responses:
        "201":
          description: Tour created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/tours/{id}:
    put:
      operationId: updateAdminTour
      tags: [admin]
      summary: Update a tour
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TourInput"
      responses:
        "200":
          description: Tour updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteAdminTour
      tags: [admin]
      summary: Delete a tour
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionReceipt"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/destinations:
    get:
      operationId: getAdminDestinations
      tags: [admin]
      summary: List all destinations for admin
      responses:
        "200":
          description: All destinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DestinationDetail"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createAdminDestination
      tags: [admin]
      summary: Create a destination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationInput"
      responses:
        "201":
          description: Destination created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationDetail"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/destinations/{id}:
    put:
      operationId: updateAdminDestination
      tags: [admin]
      summary: Update a destination
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationInput"
      responses:
        "200":
          description: Destination updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationDetail"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteAdminDestination
      tags: [admin]
      summary: Delete a destination
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionReceipt"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/bookings:
    get:
      operationId: getAdminBookings
      tags: [admin]
      summary: List all bookings
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/bookings/{id}:
    patch:
      operationId: updateAdminBookingStatus
      tags: [admin]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/inquiries:
    get:
      operationId: getAdminInquiries
      tags: [admin]
      summary: List all inquiries
      responses:
        "200":
          description: Inquiries list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InquiryItem"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/inquiries/{id}:
    patch:
      operationId: updateAdminInquiryStatus
      tags: [admin]
      summary: Update inquiry status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryStatusUpdate"
      responses:
        "200":
          description: Updated inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InquiryItem"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/blog:
    post:
      operationId: createAdminBlogPost
      tags: [admin]
      summary: Create a blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/blog/{id}:
    delete:
      operationId: deleteAdminBlogPost
      tags: [admin]
      summary: Delete a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionReceipt"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/faqs:
    post:
      operationId: createAdminFaq
      tags: [admin]
      summary: Create a FAQ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqItem"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/faqs/{id}:
    delete:
      operationId: deleteAdminFaq
      tags: [admin]
      summary: Delete a FAQ
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionReceipt"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/gallery:
    post:
      operationId: createAdminGalleryImage
      tags: [admin]
      summary: Add gallery image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryImageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/gallery/{id}:
    delete:
      operationId: deleteAdminGalleryImage
      tags: [admin]
      summary: Delete gallery image
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionReceipt"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Tour:
      type: object
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        destination: { type: string }
        country: { type: string }
        region: { type: ["string", "null"] }
        durationDays: { type: integer }
        category: { type: string }
        tourType: { type: string, enum: [private, group, flexible] }
        priceAmount: { type: ["number", "null"] }
        priceCurrency: { type: ["string", "null"] }
        availabilityStatus: { type: string, enum: [inquiry_only, scheduled, unavailable] }
        imageUrl: { type: string }
        summary: { type: string }
        description: { type: string }
        highlights:
          type: array
          items: { type: string }
        itinerary:
          type: array
          items:
            $ref: "#/components/schemas/ItineraryDay"
        included:
          type: array
          items: { type: string }
        notIncluded:
          type: array
          items: { type: string }
        isFeatured: { type: boolean }
        isDemo: { type: boolean }
        isPublished: { type: boolean }
      required:
        - id
        - slug
        - title
        - destination
        - country
        - region
        - durationDays
        - category
        - tourType
        - priceAmount
        - priceCurrency
        - availabilityStatus
        - imageUrl
        - summary
        - description
        - highlights
        - itinerary
        - included
        - notIncluded
        - isFeatured
        - isDemo
...
    TourInput:
      type: object
      properties:
        slug: { type: string, minLength: 2, maxLength: 100 }
        title: { type: string, minLength: 3, maxLength: 200 }
        destinationId: { type: integer }
        durationDays: { type: integer, minimum: 1 }
        category: { type: string, minLength: 2 }
        tourType: { type: string, enum: [private, group, flexible] }
        priceAmount: { type: ["number", "null"] }
        priceCurrency: { type: ["string", "null"] }
        availabilityStatus: { type: string, enum: [inquiry_only, scheduled, unavailable] }
        imageUrl: { type: string, minLength: 1 }
        summary: { type: string, minLength: 10 }
        description: { type: string, minLength: 20 }
        highlights:
          type: array
          items: { type: string }
        itinerary:
          type: array
          items:
            $ref: "#/components/schemas/ItineraryDay"
        included:
          type: array
          items: { type: string }
        notIncluded:
          type: array
          items: { type: string }
        isFeatured: { type: boolean }
        isDemo: { type: boolean }
        isPublished: { type: boolean }
      required: [slug, title, destinationId, durationDays, category, tourType, imageUrl, summary, description]
...
    DestinationDetail:
      allOf:
        - $ref: "#/components/schemas/Destination"
        - type: object
          properties:
            gallery:
              type: array
              items: { type: string }
            bestTimeToVisit: { type: ["string", "null"] }
            languages:
              type: array
              items: { type: string }
            currencyCode: { type: ["string", "null"] }
            travelNotes: { type: ["string", "null"] }
            tourCount: { type: integer }
          required: [gallery, bestTimeToVisit, languages, currencyCode, travelNotes, tourCount]
...
    BookingStatusUpdate:
      type: object
      properties:
        status: { type: string, enum: [pending, confirmed, cancelled] }
      required: [status]
...
    InquiryStatusUpdate:
      type: object
      properties:
        status: { type: string, enum: [new, in_progress, resolved] }
      required: [status]
...
    SubmissionReceipt:
      type: object
      properties:
        id: { type: integer }
        message: { type: string }
      required: [id, message]
...